# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AfroMarket",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7000/afro-market/v1",
      "variables": {}
    },
    {
      "url": "https://afromarket-be.herokuapp.com/afro-market/v1",
      "variables": {}
    }
  ],
  "paths": {
...
    "/order/create": {
      "post": {
        "tags": [
          "order"
        ],
        "summary": "create order",
        "operationId": "createorder",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createorderrequest"
                  },
                  {
                    "example": {
                      "order_tray": [
                        {
                          "id": "7bd91bd6-1472-44d1-8447-3401c0d7bc72",
                          "quantity_ordered": 1
                        },
                        {
                          "id": "c99eb99f-c63e-4d10-b678-23d2b2abd742",
                          "quantity_ordered": 1
                        }
                      ],
                      "delivery_address": "1, afroworld africa to the world"
                    }
                  }
                ]
              },
              "example": {
                "order_tray": [
                  {
                    "id": "7bd91bd6-1472-44d1-8447-3401c0d7bc72",
                    "quantity_ordered": 1
                  },
                  {
                    "id": "c99eb99f-c63e-4d10-b678-23d2b2abd742",
                    "quantity_ordered": 1
                  }
                ],
                "delivery_address": "1, afroworld africa to the world"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
    "/order/my-orders": {
      "get": {
        "tags": [
          "order"
        ],
        "summary": "get my orders",
        "operationId": "getmyorders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
    "/inventory/remove/11": {
      "patch": {
        "tags": [
          "inventory"
        ],
        "summary": "remove product from inventory",
        "operationId": "removeproductfrominventory",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
    "/dispute/create/435f5bc4-e9a5-4078-882b-95fb610e796c": {
      "post": {
        "tags": [
          "dispute"
        ],
        "summary": "create dispute",
        "operationId": "createdispute",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createdisputerequest"
                  },
                  {
                    "example": {
                      "reason": "Wrong Order",
                      "description": "fake product was delivered"
                    }
                  }
                ]
              },
              "example": {
                "reason": "Wrong Order",
                "description": "fake product was delivered"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
    "/dispute/all": {
      "get": {
        "tags": [
          "dispute"
        ],
        "summary": "get my disputes",
        "operationId": "getmydisputes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "createorderrequest": {
        "title": "createorderrequest",
        "required": [
          "order_tray",
          "delivery_address"
        ],
        "type": "object",
        "properties": {
          "order_tray": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderTray"
            },
            "description": ""
          },
          "delivery_address": {
            "type": "string"
          }
        },
        "example": {
          "order_tray": [
            {
              "id": "7bd91bd6-1472-44d1-8447-3401c0d7bc72",
              "quantity_ordered": 1
            },
            {
              "id": "c99eb99f-c63e-4d10-b678-23d2b2abd742",
              "quantity_ordered": 1
            }
          ],
          "delivery_address": "1, afroworld africa to the world"
        }
      },
      "OrderTray": {
        "title": "OrderTray",
        "required": [
          "id",
          "quantity_ordered"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "quantity_ordered": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "id": "7bd91bd6-1472-44d1-8447-3401c0d7bc72",
          "quantity_ordered": 1
        }
      },
...
      "createdisputerequest": {
        "title": "createdisputerequest",
        "required": [
          "reason",
          "description"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "example": {
          "reason": "Wrong Order",
          "description": "fake product was delivered"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}